requirements.txt
pandas
numpy
music_tag
itunesLibrary
pytest
pillow